requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
chromadb>=0.5.0
openai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.30.0
httpx>=0.27.0
pydantic>=2.0.0
tavily-python>=0.5.0
azure-servicebus>=7.12.0
pyarrow>=15.0.0
duckdb>=1.0.0
